[TOOLS-62] Separate buffer services
